Changed tests to unit tests
